# backend/gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "企业级 AI 智能问答系统 - API 网关\n\n## 服务架构\n- **API 网关** (本服务) → 认证/限流/路由/熔断\n- **RAG 服务** → 文档索引/向量检索/LLM 问答\n- **PostgreSQL** → 用户/对话/审计日志\n- **Redis** → 缓存/分布式限流/Token 黑名单\n- **Milvus** → 向量数据库\n\n## 认证方式\n```\nPOST /api/v1/auth/login  →  获取 JWT Token\nAuthorization: Bearer \u003ctoken\u003e  →  访问受保护端点\n```\n\n## 数据合规\n- 数据导出: GET /api/v1/user/export (PIPL §45)\n- 删除账户: POST /api/v1/user/delete-request (PIPL §47)\n- 审计日志: 所有写操作自动记录",
        "title": "AI QA System API",
        "termsOfService": "https://github.com/nieen/ai-qa-system",
        "contact": {
            "name": "AI QA System Team",
            "url": "https://github.com/nieen/ai-qa-system"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/knowledge-bases": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库"
                ],
                "summary": "获取知识库列表",
                "responses": {
                    "200": {
                        "description": "知识库列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/knowledge-bases/{kbId}/chat": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "发送问题进行知识库问答，返回 SSE 流式响应",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "问答"
                ],
                "summary": "知识库问答",
                "parameters": [
                    {
                        "type": "string",
                        "description": "知识库 ID",
                        "name": "kbId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE 事件流",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Token (格式: Bearer \\\u003ctoken\\\u003e)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}